Split out fetchSitesAndUser
